BSP-180: refactoring
