+++ src/main/java/uk/gov/hmcts/reform/bsp/common/model/shared/CaseDetails.java
@@ -1,5 +1,6 @@
-package uk.gov.hmcts.reform.bsp.common.model.transformation.output;
+package uk.gov.hmcts.reform.bsp.common.model.shared;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.Getter;
@@ -8,24 +9,25 @@ import java.util.Map;
 
 @Builder
 @Getter
-public class CaseCreationDetails {
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class CaseDetails {
+
+    @JsonProperty("id")
+    private final String caseId;
 
     @JsonProperty("case_type_id")
     private final String caseTypeId;
 
-    @JsonProperty("event_id")
-    private final String eventId;
-
     @JsonProperty("case_data")
     private final Map<String, Object> caseData;
 
-    public CaseCreationDetails(
+    public CaseDetails(
+            String caseId,
             String caseTypeId,
-            String eventId,
             Map<String, Object> caseData
     ) {
+        this.caseId = caseId;
         this.caseTypeId = caseTypeId;
-        this.eventId = eventId;
         this.caseData = caseData;
     }
 }
