+++ src/main/java/uk/gov/hmcts/reform/bsp/common/model/update/in/BulkScanCaseUpdateRequest.java
@@ -2,7 +2,8 @@ package uk.gov.hmcts.reform.bsp.common.model.update.in;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
-import uk.gov.hmcts.reform.bsp.common.model.transformation.in.ExceptionRecord;
+import uk.gov.hmcts.reform.bsp.common.model.shared.CaseDetails;
+import uk.gov.hmcts.reform.bsp.common.model.shared.in.ExceptionRecord;
 
 @Getter
 public class BulkScanCaseUpdateRequest {
@@ -11,8 +12,8 @@ public class BulkScanCaseUpdateRequest {
     private final CaseDetails caseDetails;
 
     public BulkScanCaseUpdateRequest(
-        @JsonProperty("exception_record") ExceptionRecord exceptionRecord,
-        @JsonProperty("case_details") CaseDetails caseDetails
+            @JsonProperty("exception_record") ExceptionRecord exceptionRecord,
+            @JsonProperty("case_details") CaseDetails caseDetails
     ) {
         this.exceptionRecord = exceptionRecord;
         this.caseDetails = caseDetails;
