+++ src/main/java/uk/gov/hmcts/reform/bsp/common/model/update/output/SuccessfulUpdateResponse.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.reform.bsp.common.model.update.output;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.Data;
-import uk.gov.hmcts.reform.bsp.common.model.update.in.CaseUpdateDetails;
+import uk.gov.hmcts.reform.bsp.common.model.shared.CaseDetails;
 
 import java.util.List;
 
@@ -12,15 +12,15 @@ import java.util.List;
 public class SuccessfulUpdateResponse {
 
     @JsonProperty("case_update_details")
-    public final CaseUpdateDetails caseUpdateDetails;
+    public final CaseDetails caseDetails;
 
     @JsonProperty("warnings")
     public final List<String> warnings;
 
     public SuccessfulUpdateResponse(
-        CaseUpdateDetails caseUpdateDetails,
-        List<String> warnings) {
-        this.caseUpdateDetails = caseUpdateDetails;
+            CaseDetails caseDetails,
+            List<String> warnings) {
+        this.caseDetails = caseDetails;
         this.warnings = warnings;
     }
 }
