+++ src/main/java/uk/gov/hmcts/reform/bsp/common/model/validation/in/OcrDataValidationRequest.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.reform.bsp.common.model.validation.in;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
+import uk.gov.hmcts.reform.bsp.common.model.shared.in.OcrDataField;
 
 import javax.validation.constraints.NotEmpty;
 import java.util.List;
@@ -15,7 +16,7 @@ public class OcrDataValidationRequest {
     private final List<OcrDataField> ocrDataFields;
 
     public OcrDataValidationRequest(
-        @JsonProperty("ocr_data_fields") List<OcrDataField> ocrDataFields
+            @JsonProperty("ocr_data_fields") List<OcrDataField> ocrDataFields
     ) {
         this.ocrDataFields = ocrDataFields;
     }
