+++ src/main/java/uk/gov/hmcts/reform/bsp/common/service/PostcodeValidator.java
@@ -10,15 +10,15 @@ public class PostcodeValidator {
         List<String> validationMessages = new ArrayList<>();
         if (fieldsMap.containsKey(postcodeKey)) {
             String postcodeValue = fieldsMap.get(postcodeKey);
-            if (isPostcodeLengthInvalid(postcodeValue)) {
+            if (!isPostcodeLengthValid(postcodeValue)) {
                 validationMessages.add(postcodeKey + " is usually 6 or 7 characters long");
             }
         }
         return validationMessages;
     }
 
-    private static boolean isPostcodeLengthInvalid(String postcode) {
+    private static boolean isPostcodeLengthValid(String postcode) {
         int postcodeLength = postcode.length();
-        return postcodeLength > 8 || postcodeLength < 6;
+        return postcodeLength >= 6 && postcodeLength <= 8;
     }
 }
